login, logout, profile, registration, search_events, search_events_curlAPI moved to dynamic folder and corresponding actions moved to dynamic_controller. moved php code from login, registration, and logout to dynamic_controller as per MVC. Profile page still needs to be refactored regarding php code.
