+++ views/dynamic/login.php
@@ -1,0 +1,19 @@
+<div class="container">
+   <div id="img">
+        <img id="avatar" src="views/images/img_avatar2.png"/>
+        <h1>Log In</h1>
+    </div>
+    <form class="form-signin" method="POST">
+        <div class="container" id="container">
+            <label for="uname"><b>Username</b></label>
+            <input type="text" placeholder="Enter Username" name="emailAddress" required>
+            <label for="psw"><b>Password</b></label>
+            <input type="password" placeholder="Enter Password" name="password" required>
+            <button class="btn btn-lg btn-primary btn-block" type="submit" name="login_btn">Login</button>
+        </div>
+    </form>
+</div>
+
+
+
+
