+++ views/dynamic/logout.php
@@ -1,0 +1,6 @@
+<?php
+//
+//session_start();
+//session_destroy();
+//header('Location: ?controller=static&action=landing');
+//?>
