+++ views/dynamic/search_events.php
@@ -24,7 +24,7 @@
 
 <!--    <link rel="stylesheet" href="../styles/styles.css">-->
 
-    <script src="../scripts/main.js"></script>
+    <script src="views/scripts/main.js"></script>
     <script async defer
             src="https://maps.googleapis.com/maps/api/js?key=AIzaSyCf6m35Pbc-BmCZzxrlUlCWmMpHOrWyMms&callback=initMap">
     </script>
