--- views/static/logout.php
@@ -1,6 +1,0 @@
-<?php
-
-session_start();
-session_destroy();
-header('Location: ?controller=static&action=landing');
-?>
