+++ views/templates/header.php
@@ -7,12 +7,13 @@
             session_start();
             //3.1.4 if the user is logged in Greets the user with message
             if (isset($_SESSION['emailAddress'])) {
-                echo '<li class="nav-item"><a class="nav-link" href="?controller=static&action=profile">Profile</a></li>';
-                echo '<li class="nav-item"><a class="nav-link" href="?controller=static&action=logout">Logout</a></li>';
+                echo '<li class="nav-item"><a class="nav-link" href="?controller=dynamic&action=home">Home</a></li>';
+                echo '<li class="nav-item"><a class="nav-link" href="?controller=dynamic&action=profile">Profile</a></li>';
+                echo '<li class="nav-item"><a class="nav-link" href="?controller=dynamic&action=logout">Logout</a></li>';
             } else {
                 //3.2 When the user visits the page first time, simple login form will be displayed.
-                echo '<li class="nav-item"><a class="nav-link" href="?controller=static&action=registration">Sign Up</a></li>';
-                echo '<li class="nav-item"><a class="nav-link" href="?controller=static&action=login">Login</a></li>';
+                echo '<li class="nav-item"><a class="nav-link" href="?controller=dynamic&action=registration">Sign Up</a></li>';
+                echo '<li class="nav-item"><a class="nav-link" href="?controller=dynamic&action=login">Login</a></li>';
             }
             ?>
         </ul>
